+++ app/src/main/java/com/example/paypay/models/ConvertedCurrencyRate.kt
@@ -1,3 +1,3 @@
 package com.example.paypay.models
 
-class ConvertedCurrencyRate(val currencySymbol : String, val convertedValue : Float)
+class ConvertedCurrencyRate(val currencySymbol : String, val convertedValue : Double)
